Refactor ModulesList component and add stories for the same.
