+++ stories/modules-list.stories.js
@@ -1,0 +1,43 @@
+/**
+ * ModulesList stories.
+ *
+ * Google Site Kit, Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+/**
+ * External dependencies
+ */
+import { storiesOf } from '@storybook/react';
+
+/**
+ * Internal dependencies
+ */
+import ModulesList from '../assets/js/components/ModulesList';
+import { WithTestRegistry } from '../tests/js/utils';
+import { STORE_NAME as CORE_MODULES } from '../assets/js/googlesitekit/modules/datastore/constants';
+import FIXTURES from '../assets/js/googlesitekit/modules/datastore/fixtures';
+
+storiesOf( 'Global', module )
+	.add( 'ModulesList', () => {
+		const setupRegistry = ( { dispatch } ) => {
+			dispatch( CORE_MODULES ).receiveGetModules( FIXTURES );
+		};
+
+		return (
+			<WithTestRegistry callback={ setupRegistry } >
+				<ModulesList />
+			</WithTestRegistry>
+		);
+	} );
